requirements.txt
loguru
networkx
scipy
numpy
shapely
websocket-client
msgpack
msgpack-numpy
websockets
omegaconf
hydra-core
apollo-modules==7.0.2
pyzmq
docker
tqdm
rich
rtree
pydantic
nvitop
tabulate
opencv-python
matplotlib
natsort